+++ packages/transaction-pay-controller/src/TransactionPayController.ts
@@ -13,7 +13,6 @@ import {
 } from './constants';
 import { QuoteRefresher } from './helpers/QuoteRefresher';
 import type {
-  AccountSupports7702Callback,
   GetDelegationTransactionCallback,
   TransactionConfigCallback,
   TransactionData,
@@ -23,6 +22,7 @@ import type {
   UpdateFiatPaymentRequest,
   UpdatePaymentTokenRequest,
 } from './types';
+import { KEYRING_TYPES_SUPPORTING_7702 } from './types';
 import { getStrategyOrder } from './utils/feature-flags';
 import { updateQuotes } from './utils/quotes';
 import { updateSourceAmounts } from './utils/source-amounts';
@@ -54,8 +54,6 @@ export class TransactionPayController extends BaseController<
   TransactionPayControllerState,
   TransactionPayControllerMessenger
 > {
-  readonly #accountSupports7702: AccountSupports7702Callback;
-
   readonly #getDelegationTransaction: GetDelegationTransactionCallback;
 
   readonly #getStrategy?: (
@@ -67,7 +65,6 @@ export class TransactionPayController extends BaseController<
   ) => TransactionPayStrategy[];
 
   constructor({
-    accountSupports7702,
     getDelegationTransaction,
     getStrategy,
     getStrategies,
@@ -81,7 +78,6 @@ export class TransactionPayController extends BaseController<
       state: { ...getDefaultState(), ...state },
     });
 
-    this.#accountSupports7702 = accountSupports7702;
     this.#getDelegationTransaction = getDelegationTransaction;
     this.#getStrategy = getStrategy;
     this.#getStrategies = getStrategies;
@@ -99,7 +95,7 @@ export class TransactionPayController extends BaseController<
 
     // eslint-disable-next-line no-new
     new QuoteRefresher({
-      accountSupports7702: this.#accountSupports7702,
+      accountSupports7702: this.#accountSupports7702.bind(this),
       getStrategies: this.#getStrategiesWithFallback.bind(this),
       messenger,
       updateTransactionData: this.#updateTransactionData.bind(this),
@@ -257,7 +253,7 @@ export class TransactionPayController extends BaseController<
 
     if (shouldUpdateQuotes) {
       updateQuotes({
-        accountSupports7702: this.#accountSupports7702,
+        accountSupports7702: this.#accountSupports7702.bind(this),
         getStrategies: this.#getStrategiesWithFallback.bind(this),
         messenger: this.messenger,
         transactionData: this.state.transactionData[transactionId],
@@ -265,6 +261,17 @@ export class TransactionPayController extends BaseController<
         updateTransactionData: this.#updateTransactionData.bind(this),
       }).catch(noop);
     }
+  }
+
+  async #accountSupports7702(account: string): Promise<boolean> {
+    const { keyrings } = this.messenger.call('KeyringController:getState');
+    const keyring = keyrings.find((k) =>
+      k.accounts.some((a) => a.toLowerCase() === account.toLowerCase()),
+    );
+
+    return keyring
+      ? (KEYRING_TYPES_SUPPORTING_7702 as string[]).includes(keyring.type)
+      : true;
   }
 
   #getStrategiesWithFallback(